requirements.txt
torch==1.13.1
torchmetrics==1.3.0.post0
pytorch-lightning==2.1.3
ray[tune]
omegaconf
wandb
hydra-core
gymnasium
h5py
scikit-learn
matplotlib
python-dotenv
imageio[ffmpeg]
motornet
ipykernel
ipywidgets
opencv-python
git+https://github.com/mitchellostrow/DSA.git#egg=DSA
